+++ CHANGELOG.md
@@ -6,6 +6,46 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Releases
 
+### [5.0.34]
+
+#### Added
+ - NEW
+
+#### Fixed
+ - [PPC. Fix Payment Methods Pro support for PayPal Commerce (#1696)](https://github.com/eventespresso/cafe/pull/1696)
+ - [Fix Conflict With REM Styles (#1697)](https://github.com/eventespresso/cafe/pull/1697)
+ - [Fix registration question admin label field not saving (#1647)](https://github.com/eventespresso/cafe/pull/1647)
+ - FIX
+
+#### Changed
+ - MOD
+
+#### Deprecated
+ - DEP
+
+#### Removed
+ - RMV
+
+#### Security
+ - SEC
+
+
+
+### [..]
+
+
+#### Fixed
+ - [Fix SaaS Signup Form (#1660)](https://github.com/eventespresso/cafe/pull/1660)
+ - [Fix SPCO Attendee Info Ticket Display (#1636)](https://github.com/eventespresso/cafe/pull/1636)
+
+#### Changed
+ - [BuildMachine 5.0.33 changes (#1683)](https://github.com/eventespresso/cafe/pull/1683)
+
+
+
+
+
+
 ### [5.0.33]
 
 #### Fixed